rename debugging manager and add to web service registry
